+++ source/Internal/Framework/Module/Configuration/DataMapper/ModuleConfigurationDataMapper.php
@@ -31,6 +31,7 @@ class ModuleConfigurationDataMapper implements ModuleConfigurationDataMapperInte
         $data = [
             'id' => $configuration->getId(),
             'path' => $configuration->getPath(),
+            'sourcePath' => $configuration->getSourcePath(),
             'version' => $configuration->getVersion(),
             'configured' => $configuration->isConfigured(),
             'title' => $configuration->getTitle(),
@@ -60,6 +61,7 @@ class ModuleConfigurationDataMapper implements ModuleConfigurationDataMapperInte
         $moduleConfiguration
             ->setId($data['id'])
             ->setPath($data['path'])
+            ->setSourcePath($data['sourcePath'])
             ->setVersion($data['version'])
             ->setConfigured($data['configured'])
             ->setTitle($data['title']);