+++ tests/Unit/Application/Controller/Admin/ModuleListTest.php
@@ -72,12 +72,14 @@ class ModuleListTest extends \OxidTestCase
         $moduleA
             ->setId('a')
             ->setPath('some')
+            ->setSourcePath('some')
             ->setTitle(['en' => 'A']);
 
         $moduleB = new ModuleConfiguration();
         $moduleB
             ->setId('b')
             ->setPath('some')
+            ->setSourcePath('some')
             ->setTitle(['en' => 'B']);
 
         $shopConfiguration->addModuleConfiguration($moduleB);